feat: update config follow official guide
