+++ apps/expo/app.json
@@ -3,7 +3,7 @@
     "name": "yourprojectsname",
     "slug": "yourprojectsname",
     "scheme": "yourprojectsname",
-    "version": "0.2.6",
+    "version": "0.2.7",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "automatic",
