--- apps/expo/app/tamaguiFonts.native.ts
@@ -1,21 +1,0 @@
-export const tamaguiFonts = {
-  InterBlack: require('@tamagui/font-inter/otf/Inter-Black.otf'),
-  InterBlackItalic: require('@tamagui/font-inter/otf/Inter-BlackItalic.otf'),
-  InterBold: require('@tamagui/font-inter/otf/Inter-Bold.otf'),
-  InterBoldItalic: require('@tamagui/font-inter/otf/Inter-BoldItalic.otf'),
-  InterExtraBold: require('@tamagui/font-inter/otf/Inter-ExtraBold.otf'),
-  InterExtraBoldItalic: require('@tamagui/font-inter/otf/Inter-ExtraBoldItalic.otf'),
-  InterExtraLight: require('@tamagui/font-inter/otf/Inter-ExtraLight.otf'),
-  InterExtraLightItalic: require('@tamagui/font-inter/otf/Inter-ExtraLightItalic.otf'),
-  InterItalic: require('@tamagui/font-inter/otf/Inter-Italic.otf'),
-  InterLight: require('@tamagui/font-inter/otf/Inter-Light.otf'),
-  InterLightItalic: require('@tamagui/font-inter/otf/Inter-LightItalic.otf'),
-  InterMedium: require('@tamagui/font-inter/otf/Inter-Medium.otf'),
-  InterMediumItalic: require('@tamagui/font-inter/otf/Inter-MediumItalic.otf'),
-  InterSemiBold: require('@tamagui/font-inter/otf/Inter-SemiBold.otf'),
-  InterSemiBoldItalic: require('@tamagui/font-inter/otf/Inter-SemiBoldItalic.otf'),
-  InterThin: require('@tamagui/font-inter/otf/Inter-Thin.otf'),
-  InterThinItalic: require('@tamagui/font-inter/otf/Inter-ThinItalic.otf'),
-  Silkscreen: require('@tamagui/font-silkscreen/files/slkscr.ttf'),
-  SilkscreenBold: require('@tamagui/font-silkscreen/files/slkscrb.ttf'),
-}
