+++ apps/expo/metro.config.js
@@ -1,5 +1,8 @@
 // Learn more https://docs.expo.dev/guides/monorepos
 // Learn more https://docs.expo.io/guides/customizing-metro
+/**
+ * @type {import('expo/metro-config')}
+ */
 const { getDefaultConfig } = require('expo/metro-config')
 const path = require('path')
 
@@ -8,11 +11,7 @@ const projectRoot = __dirname
 // This can be replaced with `find-yarn-workspace-root`
 const workspaceRoot = path.resolve(projectRoot, '../..')
 
-/** @type {import('expo/metro-config').MetroConfig} */
-const config = getDefaultConfig(projectRoot, {
-  // [Web-only]: Enables CSS support in Metro.
-  isCSSEnabled: true,
-})
+const config = getDefaultConfig(projectRoot)
 
 // 1. Watch all files within the monorepo
 config.watchFolders = [workspaceRoot]
@@ -21,8 +20,6 @@ config.resolver.nodeModulesPaths = [
   path.resolve(projectRoot, 'node_modules'),
   path.resolve(workspaceRoot, 'node_modules'),
 ]
-// 3. Force Metro to resolve (sub)dependencies only from the `nodeModulesPaths`
-config.resolver.disableHierarchicalLookup = true
 
 // Expo 49 issue: default metro config needs to include "mjs"
 // https://github.com/expo/expo/issues/23180
