+++ apps/expo/tsconfig.json
@@ -1,6 +1,9 @@
 {
-  "extends": "../../tsconfig",
+  "extends": "../../tsconfig.base",
+  "include": ["app/**/*.ts", "app/**/*.tsx"],
   "compilerOptions": {
+    "composite": true,
     "noEmit": true
-  }
+  },
+  "references": []
 }
