+++ apps/next/types.d.ts
@@ -1,14 +1,12 @@
 // don't import from here, that's handled already
 // instead this is just setting types for this folder
 
-import { config } from '@my/ui'
+import { config } from '@my/config'
 
-type AppConfig = typeof config
+export type AppConfig = typeof config
 
-declare module 'tamagui' {
+declare module '@my/ui' {
   // overrides TamaguiCustomConfig so your custom types
   // work everywhere you import `tamagui`
   interface TamaguiCustomConfig extends AppConfig {}
 }
-
-export default config
