+++ packages/app/features/home/screen.tsx
@@ -20,7 +20,7 @@ export function HomeScreen() {
 
   return (
     <YStack f={1} jc="center" ai="center" p="$4" space>
-      <YStack space="$4" maw={600} jc="center">
+      <YStack space="$4" bc="$" jc="center">
         <Clock />
         <H1 ta="center" ff="$silkscreen">
           Welcome to Tamagui.
