+++ packages/app/provider/Provider.web.tsx
@@ -1,6 +1,5 @@
-import { CustomToast, TamaguiProvider, TamaguiProviderProps, ToastProvider } from '@my/ui'
+import { CustomToast, TamaguiProvider, TamaguiProviderProps, ToastProvider, config } from '@my/ui'
 import { NextThemeProvider, useRootTheme, useThemeSetting } from '@tamagui/next-theme'
-import config from 'app/tamagui.config'
 import { createThemeStore, type mode, useThemeStore } from 'app/zustand'
 import { useEffect } from 'react'
 import { ToastViewport } from './ToastViewport'
