+++ packages/config/.gitignore
@@ -1,0 +1,5 @@
+dist/
+.DS_Store
+THUMBS_DB
+node_modules/
+types/
