+++ packages/config/package.json
@@ -1,0 +1,32 @@
+{
+    "name": "@my/config",
+    "version": "0.2.7",
+    "private": true,
+    "main": "./src/index.tsx",
+    "types": "./src",
+    "files": [
+        "types",
+        "dist"
+    ],
+    "sideEffects": [
+        "*.css"
+    ],
+    "scripts": {
+        "config:build": "tamagui-build --skip-types",
+        "config:watch": "tamagui-build --skip-types --watch"
+    },
+    "dependencies": {
+        "@my/ui": "*",
+        "@tamagui/animations-moti": "latest",
+        "@tamagui/font-inter": "latest",
+        "@tamagui/font-silkscreen": "latest",
+        "@tamagui/react-native-media-driver": "latest",
+        "@tamagui/shorthands": "latest",
+        "@tamagui/themes": "latest",
+        "react-native-reanimated": "~3.3.0",
+        "tamagui": "latest"
+    },
+    "devDependencies": {
+        "@tamagui/build": "latest"
+    }
+}
