+++ packages/config/src/animations.ts
