+++ packages/config/src/tamagui.config.ts
@@ -1,8 +1,9 @@
 import { shorthands } from '@tamagui/shorthands'
-import { themes, tokens } from '@tamagui/themes'
+import { tokens } from '@tamagui/themes/v2'
+import { themes } from '@tamagui/themes/v2-themes'
 import { createTamagui } from 'tamagui'
 import { animations } from './animations'
-import fonts from './fonts'
+import { fonts } from './fonts'
 import { media } from './mediaQueries'
 
 export const config = createTamagui({
@@ -10,9 +11,16 @@ export const config = createTamagui({
   animations,
   shouldAddPrefersColorThemes: true,
   themeClassNameOnRoot: false,
+  onlyAllowShorthands: true,
   shorthands,
   fonts,
   themes,
   tokens,
   media,
+  settings: {
+    allowedStyleValues: 'somewhat-strict',
+  },
 })
+
+// for the compiler to find it
+export default config
