+++ packages/ui/src/NativeToast.tsx
@@ -20,8 +20,8 @@ export const NativeToast = () => {
       scale={1}
       animation="quick"
     >
-      <YStack>
-        <Toast.Title>{currentToast.title}</Toast.Title>
+      <YStack px="$2" py="$1.5">
+        <Toast.Title lh="$1">{currentToast.title}</Toast.Title>
         {!!currentToast.message && <Toast.Description>{currentToast.message}</Toast.Description>}
       </YStack>
     </Toast>
