+++ packages/ui/src/index.tsx
@@ -1,5 +1,5 @@
 export * from 'tamagui'
 export * from '@tamagui/toast'
-export { config } from './tamagui.config'
+export { config } from '@my/config'
 export * from './CustomToast'
 export * from './MyComponent'
