+++ tsconfig.json
@@ -5,12 +5,13 @@
     "noUncheckedIndexedAccess": true,
     "paths": {
       "app/*": ["./packages/app/*"],
+      "@my/config/*": ["./packages/config/*"],
       "@my/ui/*": ["./packages/ui/*"]
     },
     "strict": true,
     "strictNullChecks": true
   },
-  "extends": "tsconfig.base",
+  "extends": "./tsconfig.base",
   "exclude": [
     "**/node_modules",
     "**/dist",
